added deleteAllTodo button

+++ frontend/src/components/ToDoList/ToDoList.jsx
@@ -2,41 +2,30 @@
 import './ToDoList.css'
 import { useSelector, useDispatch } from 'react-redux'
 import { deleteCurrentToDo } from '../../store/TodoSlice'
-import { useEffect } from 'react';
 
 const ToDoList = () => {
 
-    const {data, status,deleteStatus} = useSelector((state) => state.todoSlice);
-    // console.log(data);
-    // console.log(status);
+    const {data, status, deleteStatus} = useSelector((state) => state.todoSlice);
+
     const dispatch = useDispatch();
 
     const handleDelete = (id) => {
         dispatch(deleteCurrentToDo(id))
     }
-    
-
-    // const handleDelete = (id) => {
-    //     // console.log(id);
-    //     fetch(`http://localhost:8080/todo/${id}`,{
-    //         method: 'DELETE'
-    //     })
-    //     .then(res => res.json())
-    //     .then(result => dispatch(deleteToDo(id)))  
-    // }
    
     return (
         <div>
-        
-
-            {data?.map((item, index) => {
-                return (
-                    <div key={item._id} className="individualTask">
-                        <span className='count'>{index + 1}.</span>
-                        {item.task}
-                        <button onClick={() => handleDelete(item._id)} className='deletebtn'>Delete</button>
-                    </div>
-                )    
+            { 
+            data?.map((item, index) => {
+                if( deleteStatus == 'pending' || deleteStatus == 'rejected' ){
+                        return (
+                            <div key={item._id} className="individualTask">
+                                <span className='count'>{index + 1}.</span>
+                                <div className='todotext'>{item.task}</div>    
+                                <button onClick={() => handleDelete(item._id)} className='deletebtn'>Delete</button>
+                            </div>
+                        )     
+                }
             })}
         </div>
     )
